pyproject.toml
[project]
name = "ask-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.34.0",
    "asyncmy>=0.2.10",
    "pyjwt>=2.8.0",
    "passlib[argon2]>=1.7.4",
    "email-validator>=2.1.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    # 2FA関連
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.2",
    "google-auth>=2.23.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
]
